+++ agents/tavern/apps/web/next.config.js
@@ -19,6 +19,10 @@ const config = {
   /** We already do linting and typechecking as separate tasks in CI */
   eslint: { ignoreDuringBuilds: true },
   typescript: { ignoreBuildErrors: true },
+
+  experimental: {
+    largePageDataBytes: 10 * 1000000,
+  },
 }
 
 export default config