+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ownxai/web
 
+## 0.1.7
+
+### Patch Changes
+
+- @ownxai/api@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes