+++ packages/tokenizer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ownxai/tokenizer
 
+## 0.1.3
+
+### Patch Changes
+
+- Includes assets directory
+
 ## 0.1.2
 
 ### Patch Changes